pull out duplicated webview helpers into base class
